change integer filed to float

+++ school_diary/diary/models.py
@@ -195,7 +195,7 @@ class Administrators(models.Model):
 
 class Controls(models.Model):
     name = models.CharField(max_length=120, verbose_name='Вид работы')
-    weight = models.IntegerField(verbose_name=' Коэффицент', default=1)
+    weight = models.FloatField(verbose_name=' Коэффицент', default=1)
 
     class Meta:
         verbose_name = "Вид работы"
